refactor: move GNB struct and methods to separate file
